added a search bar on home index that allows user to search for a movie to rate by title. i have the method in place, but still need the code for, adding the movie that the api brings up into the media database.
created 'SingleRate' home/view page, which is where the search bar submit button brings the user to. 
this page shows the user what movie they selected, and will let them rate the movie.
this page eventually should have a rate more page, with a search bar (for title) that redirects to the same page